customer_service_chatbot_LLM/requirements.txt
langchain==0.0.339
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf~=3.19.0
sentence-transformers==2.2.2

# Core LLM & Embeddings
google-generativeai>=0.2.0

# Dynamic Updater
schedule>=1.1.0
pyyaml>=6.0

# Multimodal (Optional but recommended)
pillow>=9.0.0
pytesseract>=0.3.10
transformers>=4.20.0
torch>=1.13.0

# Medical QA
spacy>=3.0.0

# Research Assistant
pandas>=1.3.0
arxiv>=1.4.0

# Sentiment Analysis
vaderSentiment>=3.3.2

# Multilingual Support
langdetect>=1.0.9
googletrans==4.0.0-rc1

# Visualization & NLP
networkx>=2.6
plotly>=5.0.0

# Optional: Additional transformers models
sentencepiece>=0.1.99
